requirements.txt
fastapi==0.111.0
razorpay>=1.4.1
boto3>=1.34.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai>=1.30.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv~=1.2.2
httpx>=0.27.0
itsdangerous>=2.2.0
openai-whisper>=20231117